% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misalignDetect.R
\name{exclusiveGroups}
\alias{exclusiveGroups}
\title{Find the possible misaligned peak groups}
\usage{
exclusiveGroups(groupInfo, ppm, qThre)
}
\arguments{
\item{groupInfo}{An output list from \code{\link{splitGroupPval}} a larger
bw, usually the maximum expected retension time drift.}

\item{ppm}{Should set as same as the one used in the corresponding
\code{\link[xcms]{xcmsSet-class}} object.}

\item{qThre}{The threshould of the adjusted p-value of output peak groups.}
}
\value{
A list containing the following components:

  \item{excluGroups}{A \code{\link[xcms]{xcmsSet-class}}$\code{group}-like matrix with all detected
  misaligned peak groups.}

  \item{excluPval}{The adjusted p-values of the groups in \code{excluGroups}.}

  \item{excluLargeSmall}{The corresponding group indexes of the groups in \code{excluGroups}.}
}
\description{
This function finds possible misaligned peak groups with a thresshold of the
adjusted p-value.
}
\details{
This function finds the groups in \code{largeWinGroups} which has
  corrsponding exclusive peak groups in \code{smallWinGroups}, and output the
  groups with adjusted p-value smaller than \code{qThre}.
}
\examples{
# obtain data
data('xcmsExamples')
xcmsLargeWin <- xcmsExamples$xcmsLargeWin
xcmsSmallWin <- xcmsExamples$xcmsSmallWin
ppm <- xcmsExamples$ppm

# calculate the p-value of each peak group
groupInfo <- splitGroupPval(xcmsLargeWin, xcmsSmallWin)

# output the possible misaligned peak groups
# excluInfo <- exclusiveGroups(groupInfo, ppm, 0.05)
}
