% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/XCMSedited.R
\name{fillPeaksChromPar}
\alias{fillPeaksChromPar}
\title{Edited XCMS fillPeaksChromPar for feature-wise warping functions}
\usage{
fillPeaksChromPar(arg)
}
\arguments{
\item{arg}{A list sent from \code{\link[xcms]{fillPeaks.chrom-methods}}.}
}
\value{
A list of sameple index vector and filled peak matrix.
}
\description{
This function is edited from \code{fillPeaksChromPar} in
\code{\link[xcms]{fillPeaks.chrom-methods}} to accept feature-wise warping
functions.
}
\details{
This function is for parallelly filling missing peaks with
feature-wise warping functions. The original function in
\code{\link[xcms]{fillPeaks.chrom-methods}} can only handle sample-wise
warping functions.
}
