% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/XCMSedited.R
\name{getPeaksncGTW}
\alias{getPeaksncGTW}
\title{Edited XCMS getPeaks for feature-wise warping functions}
\usage{
getPeaksncGTW(object, peakrange, step = 0.1, naidx)
}
\arguments{
\item{object}{An \code{\link[xcms]{xcmsRaw-class}} object.}

\item{peakrange}{\code{matrix} with 4 required columns \code{"mzmin"},
\code{"mzmax"}, \code{"rtmin"} and \code{"rtmax"}.}

\item{step}{\code{numeric(1)} defining the bin size for the profile matrix
generation.}

\item{naidx}{A vector contains the sample indexes need to be filled.}
}
\value{
A list of sameple index vector and filled peak matrix.
}
\description{
This function is edited from \code{\link[xcms]{getPeaks-methods}} to accept
feature-wise warping functions.
}
\details{
This function is for parallelly filling missing peaks with
feature-wise warping functions. The original code function in
\code{\link[xcms]{getPeaks-methods}} can only handle sample-wise warping
functions.
}
