% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misalignDetect.R
\name{misalignDetect}
\alias{misalignDetect}
\title{Detect misaligned peak groups in xcmsSet object of XCMS}
\usage{
misalignDetect(xcmsLargeWin, xcmsSmallWin, ppm, qThre = 0.05,
  maxRtWin = 50)
}
\arguments{
\item{xcmsLargeWin}{A \code{\link[xcms::xcmsSet-class]{xcmsSet}} object with
a larger bw, usually the maximum expected retension time drift.}

\item{xcmsSmallWin}{A \code{\link[xcms::xcmsSet-class]{xcmsSet}}
object with a smaller bw, usually the resolution of the retension time.}

\item{ppm}{Should set as same as the one in the input
\code{\link[xcms::xcmsSet-class]{xcmsSet}} object.}

\item{qThre}{The threshould of the p-value after multiple test correction.
The default is 0.05.}

\item{maxRtWin}{The threshould of the maximum retension time range. This is
for filtering out some bad groups. The default is 50 (seconds).}
}
\value{
A \code{\link[xcms::xcmsSet-class]{xcmsSet}} object with all
detected misaligned peak groups.
}
\description{
This function detects the misaligned peak groups with two objects of class
\code{\link[xcms::xcmsSet-class]{xcmsSet}} with two different values of "bw"
parameter in \code{\link[xcms]{group}}.
}
\details{
This function includes two major steps to determine a peak group is
misaligned or not.
}
\examples{
add(1, 1)
add(10, 1)
}
