% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ncGTW.R
\docType{package}
\name{ncGTW-package}
\alias{ncGTW-package}
\alias{ncGTW}
\title{ncGTW: A package for detecting and aligning the misaligned features in
LC-MS data}
\description{
The purpose of ncGTW is to help XCMS for LC-MS data alignment. Currently,
ncGTW can detect the misaligned feature groups by XCMS, and the user can
choose to realign these feature groups by ncGTW or not.
}
\references{
Wu, Chiung-Ting, et al. Alignment of LC-MS Profiles by
Neighbor-wise Compound-specific Graphical Time Warping with Misalignment
Detection. bioRxiv, 2019, 715334.
}
