% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ncGTW.R
\name{ncGTW}
\alias{ncGTW}
\title{Run ncGTW alignment}
\usage{
ncGTW(ncGTWinput, xcmsLargeWin, parSamp = 10, bpParam = bpparam(),
  ncGTWparam = NULL)
}
\arguments{
\item{xcmsLargeWin}{A \code{\link[xcms::xcmsSet-class]{xcmsSet}} object with
a larger bw, usually the maximum expected retension time drift.}

\item{xcmsSmallWin}{A \code{\link[xcms::xcmsSet-class]{xcmsSet}}
object with a smaller bw, usually the resolution of the retension time.}
}
\value{
A \code{\link[xcms::xcmsSet-class]{xcmsSet}} object with all
detected misaligned peak groups.
}
\description{
This function calculates the p-value of each peak group in the
\code{\link[xcms::xcmsSet-class]{xcmsSet}} with the smaller "bw" parameter,
and finds the corresponding peak group in the
\code{\link[xcms::xcmsSet-class]{xcmsSet}} with the larger "bw" parameter.
}
\details{
This function includes two major steps to determine a peak group is
misaligned or not.
}
\examples{
add(1, 1)
add(10, 1)
}
