% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ncGTWalign.R
\name{ncGTWalign}
\alias{ncGTWalign}
\title{Run ncGTW alignment}
\usage{
ncGTWalign(ncGTWinput, xcmsLargeWin, parSamp = 10,
  bpParam = BiocParallel::SnowParam(workers = 1), ncGTWparam = NULL)
}
\arguments{
\item{ncGTWinput}{A  object with
a larger bw, usually the maximum expected retension time drift.}

\item{xcmsLargeWin}{A
object with a smaller bw, usually the resolution of the retension time.}

\item{parSamp}{A  object with
a larger bw, usually the maximum expected retension time drift.}

\item{bpParam}{A
object with a smaller bw, usually the resolution of the retension time.}

\item{ncGTWparam}{A
object with a smaller bw, usually the resolution of the retension time.}
}
\value{
A  object with all
detected misaligned peak groups.
}
\description{
This function calculates the p-value of each peak group in the
 with the smaller "bw" parameter,
and finds the corresponding peak group in the
 with the larger "bw" parameter.
}
\details{
This function includes two major steps to determine a peak group is
misaligned or not.
}
\examples{
# obtain data
data('xcmsExamples')
xcmsLargeWin <- xcmsExamples$xcmsLargeWin
xcmsSmallWin <- xcmsExamples$xcmsSmallWin
ppm <- xcmsExamples$ppm

# detect misaligned features
excluGroups <- misalignDetect(xcmsLargeWin, xcmsSmallWin, ppm)

# obtain the paths of the sample files
filepath <- system.file("extdata", package = "ncGTW")
file <- list.files(filepath, pattern="mzxml", full.names=TRUE)

tempInd <- matrix(0, length(file), 1)
for (n in 1:length(file)){
  tempCha <- file[n]
  tempLen <- nchar(tempCha)
  tempInd[n] <- as.numeric(substr(tempCha, regexpr("example", tempCha) + 7, tempLen - 6))
}
# sort the paths by data acquisition order
file <- file[sort.int(tempInd, index.return = TRUE)$ix]

# load the sample profiles
ncGTWinputs <- loadProfile(file, excluGroups)

# run ncGTW alignment
ncGTWoutputs <- vector('list', dim(excluGroups)[1])
for (n in 1:dim(excluGroups)[1])
 ncGTWoutputs[[n]] <- ncGTWalign(ncGTWinputs[[n]], xcmsLargeWin, 5)
}
