% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allClasses.R
\docType{class}
\name{ncGTWinput-class}
\alias{ncGTWinput-class}
\alias{ncGTWinput}
\title{Class "ncGTWinput"}
\description{
An S4 class for storing the inputs of ncGTW alignment.
}
\section{Slots}{

\describe{
\item{\code{groupInfo}}{A vector of the information of the feature.}

\item{\code{profiles}}{A raw data matrix in which each row is a sample profile.}

\item{\code{rtRaw}}{A raw RT matrix in which each row is the corresponding
sample RT. in \code{profiles}.}
}}

