% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allClasses.R
\docType{class}
\name{ncGTWoutput-class}
\alias{ncGTWoutput-class}
\alias{ncGTWoutput}
\title{Class "ncGTWoutput"}
\description{
An S4 class for storing the outputs of ncGTW alignment.
}
\section{Slots}{

\describe{
\item{\code{alignData}}{A matrix in which each row is a sample profile after
downsampling.}

\item{\code{scanRange}}{A downsampled RT matrix in which each row is the
corresponding sample RT in \code{alignData}.}

\item{\code{path}}{A list of the same length as the sample number, in which
each element is a matrix of the alignment result of the corresponding
sample.}

\item{\code{downSample}}{The factor of downsampling when perform ncGTW
alignment.}
}}

