% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allClasses.R
\docType{class}
\name{ncGTWwarp-class}
\alias{ncGTWwarp-class}
\alias{ncGTWwarp}
\title{Class "ncGTWwarp"}
\description{
An S4 class for storing the realigned RT and the peaks with adjusted RT of
ncGTW alignment.
}
\section{Slots}{

\describe{
\item{\code{rtncGTW}}{A list of the same length as the sample number, in which
each element is a vector of the realigned RT of the corresponding sample.}

\item{\code{ncGTWpeaks}}{A matrix containing peak data with adjusted RT.}
}}

