% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xcmsExamples.R
\docType{data}
\name{xcmsExamples}
\alias{xcmsExamples}
\title{Examples of xcmsSet for inputs of ncGTW}
\format{A list with two \code{\link[xcms]{xcmsSet-class}} objects and one
\code{ppm} parameter.}
\usage{
data(xcmsExamples)
}
\description{
These two \code{\link[xcms]{xcmsSet-class}} objects are created from the
example dataset. They are examples for inputs of
\code{\link{misalignDetect}}.
}
