% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ncGTWwarp.R
\name{adjustRT}
\alias{adjustRT}
\title{Adjust retention time}
\usage{
adjustRT(xcmsLargeWin, ncGTWinput, ncGTWoutput, ppm)
}
\arguments{
\item{xcmsLargeWin}{A  object with
a larger bw, usually the maximum expected retension time drift.}

\item{xcmsSmallWin}{A
object with a smaller bw, usually the resolution of the retension time.}
}
\value{
A  object with all
detected misaligned peak groups.
}
\description{
This function calculates the p-value of each peak group in the
 with the smaller "bw" parameter,
and finds the corresponding peak group in the
 with the larger "bw" parameter.
}
\details{
This function includes two major steps to determine a peak group is
misaligned or not.
}
\examples{
add(1, 1)
add(10, 1)
}
